More work

+++ src/flash/geom/Rectangle.java
@@ -1,24 +1,24 @@
 package flash.geom;
 
 public class Rectangle {
-	private int width;
-	private int height;
+	private double width;
+	private double height;
 	private double x;
 	private double y;
 
-	public void setHeight(int height) {
+	public void setHeight(double height) {
 		this.height = height;
 	}
 
-	public int getWidth() {
+	public double getWidth() {
 		return width;
 	}
 
-	public void setWidth(int width) {
+	public void setWidth(double width) {
 		this.width = width;
 	}
 
-	public int getHeight() {
+	public double getHeight() {
 		return height;
 	}
